Deprecates chifra names --named and other small cleanups
