+++ docs/content/api/openapi.yaml
@@ -640,14 +640,6 @@ paths:
           explode: true
           schema:
             type: boolean
-        - name: named
-          description: include well know token and airdrop addresses in the search
-          required: false
-          style: form
-          in: query
-          explode: true
-          schema:
-            type: boolean
         - name: addr
           description: >
             display only addresses in the results (useful for scripting, assumes --no_header)
@@ -659,6 +651,14 @@ paths:
             type: boolean
         - name: tags
           description: export the list of tags and subtags only
+          required: false
+          style: form
+          in: query
+          explode: true
+          schema:
+            type: boolean
+        - name: named
+          description: please use the --all option instead
           required: false
           style: form
           in: query
