+++ docs/content/chifra/accounts.md
@@ -254,7 +254,6 @@ Flags:
   -l, --all          include all (including custom) names in the search
   -c, --custom       include only custom named account in the search
   -p, --prefund      include prefund accounts in the search
-  -n, --named        include well know token and airdrop addresses in the search
   -a, --addr         display only addresses in the results (useful for scripting, assumes --no_header)
   -g, --tags         export the list of tags and subtags only
   -x, --fmt string   export format, one of [none|json*|txt|csv]
