+++ test/gold/tools/getBlocks/api_tests/getBlocks_deprecated_trace.txt
@@ -1,6 +1,6 @@
 blocks?blocks=2222222&trace
 {
   "errors": [
-    "chifra blocks - the --trace option is deprecated, please use traces option instead"
+    "chifra blocks - the --trace option is deprecated, please use --traces option instead"
   ]
 }
