+++ test/gold/tools/getTrans/api_tests/getTrans_deprecate_trace_1.txt
@@ -1,6 +1,6 @@
 transactions?transactions=4000002.5&trace
 {
   "errors": [
-    "chifra transactions - the --trace option is deprecated, please use traces option instead"
+    "chifra transactions - the --trace option is deprecated, please use the --traces option instead"
   ]
 }
